Moved modules around
Point and Pattern are now internal
